pyproject.toml
[project]
name = "rag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "faiss-cpu>=1.10.0",
    "google-genai>=1.12.1",
    "lxml[html-clean]>=5.4.0",
    "markitdown[all]>=0.1.1",
    "mcp[cli]>=1.6.0",
    "newspaper3k>=0.2.8",
    "numpy>=2.2.5",
    "pillow>=11.2.1",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]
